requirements.txt
# python=3.11.7
ultralytics
opencv-python
matplotlib
torchvision
tqdm
pandas